requirements.txt
Django==2.2.4
django-filter==2.2.0
django-pagination-bootstrap==1.3
django-widget-tweaks==1.4.5
geographiclib==1.49
geopy==1.20.0
pytz==2019.2
sqlparse==0.3.0
